Move ComplexityConverter to it's own file.
